added base_network; updated load model
